requirements.txt
python-docx
tk
Pillow
pyinstaller
